// src/lib/generated/billing/models/createDynamicPackageConfigDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * kaizen-billing-and-subscription
 * Kaizen Platform billing and subscription service
 * OpenAPI spec version: 1.0.0
 */
import type { CreateDynamicPackageConfigDtoValidityTimeUnit } from './createDynamicPackageConfigDtoValidityTimeUnit';

export interface CreateDynamicPackageConfigDto {
  code: string;
  name: string;
  description: string;
  /** @nullable */
  discountId?: number | null;
  currencyId: number;
  /** @nullable */
  validity?: number | null;
  /** @nullable */
  validityTimeUnit?: CreateDynamicPackageConfigDtoValidityTimeUnit;
  offersRequired: number;
}