// src/lib/generated/billing/models/updateDynamicPackageConfigDto.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * kaizen-billing-and-subscription
 * Kaizen Platform billing and subscription service
 * OpenAPI spec version: 1.0.0
 */
import type { UpdateDynamicPackageConfigDtoValidityTimeUnit } from './updateDynamicPackageConfigDtoValidityTimeUnit';

export interface UpdateDynamicPackageConfigDto {
  code?: string;
  name?: string;
  description?: string;
  /** @nullable */
  discountId?: number | null;
  /** @nullable */
  validity?: number | null;
  /** @nullable */
  validityTimeUnit?: UpdateDynamicPackageConfigDtoValidityTimeUnit;
  offersRequired?: number;
}